Backend/requirements.txt
asgiref==3.7.2
boto3==1.33.8
botocore==1.33.8
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
elastic-transport==8.10.0
idna==3.6
install==1.3.5
jmespath==1.0.1
joblib==1.3.2
lightfm==1.17
mysqlclient==2.2.0
numpy==1.26.2
pandas==2.1.3
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.8.2
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
